rename: pannel을 panel로 이름 변경
