+++ src/components/searchpage/map-panel.tsx
